FDG-6881 refactored to remove DeficitKeyTakeaways component from national-deficit-helper.js to deficit-key-takeaways.jsx
