+++ src/layouts/explainer/explainer-helpers/national-deficit/national-deficit-helper.js
@@ -1,7 +1,5 @@
-import {faChartColumn, faCoins, faHandHoldingDollar} from "@fortawesome/free-solid-svg-icons";
 import CustomLink from "../../../../components/links/custom-link/custom-link";
 import React from "react";
-import reactStringReplace from "react-string-replace";
 
 export const ChartPlaceholder = () => (
   <div
@@ -18,40 +16,6 @@ export const ChartPlaceholder = () => (
     Placeholder
   </div>
 );
-
-const thirdTakeawayText = `To pay for government programs while operating under a deficit, the federal
-    government borrows money by selling U.S. Treasury bonds, bills, and other securities.
-    The national debt is the accumulation of this borrowing along with associated interest
-    owed to investors who purchased these securities.`;
-
-const thirdTakeawayTextWithGlossaryTerm = reactStringReplace(
-  thirdTakeawayText,
-  'national debt',
-  match => {
-    return (
-      <CustomLink url={"/americas-finance-guide/national-debt/"}>
-        {match}
-      </CustomLink>
-    );
-  }
-);
-
-export const deficitKeyTakeaways = [
-  {
-    text: `A budget deficit occurs when the money going out exceeds the money coming in for a
-    given period. On this page, we calculate the deficit by the government’s fiscal year.`,
-    icon: faChartColumn
-  },
-  {
-    text: `In the last 50 years, the federal government budget has run a surplus five times,
-    most recently in 2001.`,
-    icon: faCoins
-  },
-  {
-    text: thirdTakeawayTextWithGlossaryTerm,
-    icon: faHandHoldingDollar
-  }
-];
 
 export const deficitLearnMoreLinks = [
   {
